250817
-배틀 행동 AI 리뉴얼 및 구현 중

+++ Assets/Scripts/UI/BattleMainUI.cs
@@ -106,17 +106,17 @@ public class BattleMainUI : UIScreen
 
     public void SetHp()
     {
-        mSlider_HP.value = PlayerManager.I.pData.HP / PlayerManager.I.pData.MaxHP * 100;
+        mSlider_HP.value = (float)PlayerManager.I.pData.HP / PlayerManager.I.pData.MaxHP * 100f;
         mTexts["GgHpTxt"].text = PlayerManager.I.pData.HP.ToString() + " / " + PlayerManager.I.pData.MaxHP.ToString();
     }
     public void SetMp()
     {
-        mSlider_MP.value = PlayerManager.I.pData.MP / PlayerManager.I.pData.MaxMP * 100;
+        mSlider_MP.value = (float)PlayerManager.I.pData.MP / PlayerManager.I.pData.MaxMP * 100;
         mTexts["GgMpTxt"].text = PlayerManager.I.pData.MP.ToString() + " / " + PlayerManager.I.pData.MaxMP.ToString();
     }
     public void SetSp()
     {
-        mSlider_SP.value = PlayerManager.I.pData.SP / PlayerManager.I.pData.MaxSP * 100;
+        mSlider_SP.value = (float)PlayerManager.I.pData.SP / PlayerManager.I.pData.MaxSP * 100;
         mTexts["GgSpTxt"].text = PlayerManager.I.pData.SP.ToString() + " / " + PlayerManager.I.pData.MaxSP.ToString();
     }
